+++ src/pages/ConsumerDebugVoucherPage/components/DebugVoucherForm.tsx
@@ -10,83 +10,6 @@ import type {
 } from '../types/debug-voucher.types'
 import { VoucherMutations } from '@/api/voucher'
 
-const mockResApi: TokenGenerationValidationResult = {
-  clientKind: 'API',
-  steps: {
-    clientAssertionValidation: {
-      result: 'PASSED',
-      failures: [],
-    },
-    publicKeyRetrieve: {
-      result: 'FAILED',
-      failures: [
-        {
-          code: '8001',
-          reason: 'I have to test a failure',
-        },
-      ],
-    },
-    clientAssertionSignatureVerification: {
-      result: 'SKIPPED',
-      failures: [],
-    },
-    platformStatesVerification: {
-      result: 'SKIPPED',
-      failures: [],
-    },
-  },
-}
-
-const mockResConsumer: TokenGenerationValidationResult = {
-  clientKind: 'CONSUMER',
-  steps: {
-    clientAssertionValidation: {
-      result: 'PASSED',
-      failures: [],
-    },
-    publicKeyRetrieve: {
-      result: 'PASSED',
-      failures: [],
-    },
-    clientAssertionSignatureVerification: {
-      result: 'PASSED',
-      failures: [],
-    },
-    platformStatesVerification: {
-      result: 'PASSED',
-      failures: [],
-    },
-  },
-  eservice: {
-    id: 'id test',
-    descriptorId: 'descriptor id test',
-    version: 'version test',
-    name: 'name test',
-  },
-}
-
-const mockResApiPassed: TokenGenerationValidationResult = {
-  clientKind: 'API',
-  steps: {
-    clientAssertionValidation: {
-      result: 'PASSED',
-      failures: [],
-    },
-    publicKeyRetrieve: {
-      result: 'PASSED',
-      failures: [],
-    },
-    clientAssertionSignatureVerification: {
-      result: 'PASSED',
-      failures: [],
-    },
-    platformStatesVerification: {
-      result: 'SKIPPED',
-      failures: [],
-    },
-  },
-}
-
 export type DebugVoucherFormValues = {
   clientAssertion: string
   clientId: string
