+++ src/pages/ConsumerDebugVoucherPage/components/DebugVoucherResultsStepsSection.tsx
@@ -3,7 +3,7 @@ import { useTranslation } from 'react-i18next'
 import { DebugVoucherResultsStep } from './DebugVoucherResultsStep'
 import { SectionContainer } from '@/components/layout/containers'
 import { Stack } from '@mui/material'
-import { useDebugVoucherContext } from './DebugVoucherContext'
+import { useDebugVoucherContext } from '../DebugVoucherContext'
 
 export const DebugVoucherResultsStepsSection: React.FC = () => {
   const { t } = useTranslation('voucher', { keyPrefix: 'consumerDebugVoucher.result.stepSection' })
